run files colletctions api

+++ src/arrayList/calc/Classificavel.java
@@ -1,5 +1,5 @@
 package arrayList.calc;
 
 public interface Classificavel {
-    int getClassificacao();
+    double getClassificacao();
 }
